eager_numba_and_floats_to_double_and_micro_optimization_erm
