+++ hierarchicalforecast/_modidx.py
@@ -129,18 +129,10 @@ d = { 'settings': { 'branch': 'main',
                                                                                                                'hierarchicalforecast/methods.py'),
                                               'hierarchicalforecast.methods._get_child_nodes': ( 'src/methods.html#_get_child_nodes',
                                                                                                  'hierarchicalforecast/methods.py'),
-                                              'hierarchicalforecast.methods._ma_cov': ( 'src/methods.html#_ma_cov',
-                                                                                        'hierarchicalforecast/methods.py'),
                                               'hierarchicalforecast.methods._reconcile_fcst_proportions': ( 'src/methods.html#_reconcile_fcst_proportions',
                                                                                                             'hierarchicalforecast/methods.py'),
-                                              'hierarchicalforecast.methods._shrunk_covariance_schaferstrimmer_no_nans': ( 'src/methods.html#_shrunk_covariance_schaferstrimmer_no_nans',
-                                                                                                                           'hierarchicalforecast/methods.py'),
-                                              'hierarchicalforecast.methods._shrunk_covariance_schaferstrimmer_with_nans': ( 'src/methods.html#_shrunk_covariance_schaferstrimmer_with_nans',
-                                                                                                                             'hierarchicalforecast/methods.py'),
                                               'hierarchicalforecast.methods.crossprod': ( 'src/methods.html#crossprod',
-                                                                                          'hierarchicalforecast/methods.py'),
-                                              'hierarchicalforecast.methods.lasso': ( 'src/methods.html#lasso',
-                                                                                      'hierarchicalforecast/methods.py')},
+                                                                                          'hierarchicalforecast/methods.py')},
             'hierarchicalforecast.probabilistic_methods': { 'hierarchicalforecast.probabilistic_methods.Bootstrap': ( 'src/probabilistic_methods.html#bootstrap',
                                                                                                                       'hierarchicalforecast/probabilistic_methods.py'),
                                                             'hierarchicalforecast.probabilistic_methods.Bootstrap.__init__': ( 'src/probabilistic_methods.html#bootstrap.__init__',
@@ -199,6 +191,13 @@ d = { 'settings': { 'branch': 'main',
                                                                                                          'hierarchicalforecast/utils.py'),
                                             'hierarchicalforecast.utils.HierarchicalPlot.plot_summing_matrix': ( 'src/utils.html#hierarchicalplot.plot_summing_matrix',
                                                                                                                  'hierarchicalforecast/utils.py'),
+                                            'hierarchicalforecast.utils._lasso': ('src/utils.html#_lasso', 'hierarchicalforecast/utils.py'),
+                                            'hierarchicalforecast.utils._ma_cov': ( 'src/utils.html#_ma_cov',
+                                                                                    'hierarchicalforecast/utils.py'),
+                                            'hierarchicalforecast.utils._shrunk_covariance_schaferstrimmer_no_nans': ( 'src/utils.html#_shrunk_covariance_schaferstrimmer_no_nans',
+                                                                                                                       'hierarchicalforecast/utils.py'),
+                                            'hierarchicalforecast.utils._shrunk_covariance_schaferstrimmer_with_nans': ( 'src/utils.html#_shrunk_covariance_schaferstrimmer_with_nans',
+                                                                                                                         'hierarchicalforecast/utils.py'),
                                             'hierarchicalforecast.utils._to_summing_matrix': ( 'src/utils.html#_to_summing_matrix',
                                                                                                'hierarchicalforecast/utils.py'),
                                             'hierarchicalforecast.utils._to_upper_hierarchy': ( 'src/utils.html#_to_upper_hierarchy',
