+++ hierarchicalforecast/utils.py
@@ -12,12 +12,20 @@ from typing import Callable, Dict, List, Optional, Iterable, Union, Sequence
 
 import matplotlib.pyplot as plt
 import numpy as np
+from numba import njit, prange
 import pandas as pd
 from sklearn.preprocessing import OneHotEncoder
 
 plt.rcParams['font.family'] = 'serif'
 
 # %% ../nbs/src/utils.ipynb 5
+# Global variables
+NUMBA_NOGIL = True
+NUMBA_CACHE = True
+NUMBA_PARALLEL = True
+NUMBA_FASTMATH = True
+
+# %% ../nbs/src/utils.ipynb 6
 class CodeTimer:
     def __init__(self, name=None, verbose=True):
         self.name = " '"  + name + "'" if name else ''
@@ -32,7 +40,7 @@ class CodeTimer:
             print('Code block' + self.name + \
                   ' took:\t{0:.5f}'.format(self.took) + ' seconds')
 
-# %% ../nbs/src/utils.ipynb 6
+# %% ../nbs/src/utils.ipynb 7
 def is_strictly_hierarchical(S: np.ndarray, 
                              tags: Dict[str, np.ndarray]):
     # main idea:
@@ -50,7 +58,7 @@ def is_strictly_hierarchical(S: np.ndarray,
     nodes = levels_.popitem()[1].size
     return paths == nodes
 
-# %% ../nbs/src/utils.ipynb 7
+# %% ../nbs/src/utils.ipynb 8
 def cov2corr(cov, return_std=False):
     """ convert covariance matrix to correlation matrix
 
@@ -69,7 +77,7 @@ def cov2corr(cov, return_std=False):
     else:
         return corr
 
-# %% ../nbs/src/utils.ipynb 9
+# %% ../nbs/src/utils.ipynb 10
 def _to_summing_matrix(S_df: pd.DataFrame, sparse_s: bool = False):
     """Transforms the DataFrame `df` of hierarchies to a summing matrix S."""
     categories = [S_df[col].unique() for col in S_df.columns]
@@ -78,9 +86,9 @@ def _to_summing_matrix(S_df: pd.DataFrame, sparse_s: bool = False):
     cats_bottom = categories[idx_bottom]
 
     try:
-        encoder = OneHotEncoder(categories=categories, sparse_output=sparse_s, dtype=np.float32)
+        encoder = OneHotEncoder(categories=categories, sparse_output=sparse_s, dtype=np.float64)
     except TypeError:  # sklearn < 1.2
-        encoder = OneHotEncoder(categories=categories, sparse=sparse_s, dtype=np.float32)
+        encoder = OneHotEncoder(categories=categories, sparse=sparse_s, dtype=np.float64)
 
     S = encoder.fit_transform(S_df).T
 
@@ -93,7 +101,7 @@ def _to_summing_matrix(S_df: pd.DataFrame, sparse_s: bool = False):
     tags = dict(zip(S_df.columns, categories))
     return S, tags
 
-# %% ../nbs/src/utils.ipynb 10
+# %% ../nbs/src/utils.ipynb 11
 def aggregate_before(df: pd.DataFrame,
               spec: List[List[str]],
               agg_fn: Callable = np.sum,
@@ -138,7 +146,7 @@ def aggregate_before(df: pd.DataFrame,
     S, tags = _to_summing_matrix(S_df.loc[bottom_hier, hiers_cols], sparse_s)
     return Y_df, S, tags
 
-# %% ../nbs/src/utils.ipynb 11
+# %% ../nbs/src/utils.ipynb 12
 def _to_upper_hierarchy(bottom_split, bottom_values, upper_key):
     upper_split = upper_key.split('/')
     upper_idxs = [bottom_split.index(i) for i in upper_split]
@@ -149,7 +157,7 @@ def _to_upper_hierarchy(bottom_split, bottom_values, upper_key):
 
     return [join_upper(val) for val in bottom_values]
 
-# %% ../nbs/src/utils.ipynb 12
+# %% ../nbs/src/utils.ipynb 13
 def aggregate(
     df: pd.DataFrame,
     spec: List[List[str]],
@@ -248,9 +256,9 @@ def aggregate(
     S[:, -1] = tags[bottom_key]
     categories = list(tags.values())
     try:
-        encoder = OneHotEncoder(categories=categories, sparse_output=sparse_s, dtype=np.float32)
+        encoder = OneHotEncoder(categories=categories, sparse_output=sparse_s, dtype=np.float64)
     except TypeError:  # sklearn < 1.2
-        encoder = OneHotEncoder(categories=categories, sparse=sparse_s, dtype=np.float32)    
+        encoder = OneHotEncoder(categories=categories, sparse=sparse_s, dtype=np.float64)    
     S = encoder.fit_transform(S).T
     if sparse_s:
         df_constructor = pd.DataFrame.sparse.from_spmatrix
@@ -259,7 +267,7 @@ def aggregate(
     S_df = df_constructor(S, index=np.hstack(categories), columns=bottom_levels)
     return Y_df, S_df, tags
 
-# %% ../nbs/src/utils.ipynb 21
+# %% ../nbs/src/utils.ipynb 22
 class HierarchicalPlot:
     """ Hierarchical Plot
 
@@ -456,7 +464,7 @@ class HierarchicalPlot:
         plt.grid()
         plt.show()
 
-# %% ../nbs/src/utils.ipynb 36
+# %% ../nbs/src/utils.ipynb 37
 # convert levels to output quantile names
 def level_to_outputs(level:Iterable[int]):
     """ Converts list of levels into output names matching StatsForecast and NeuralForecast methods.
@@ -500,7 +508,7 @@ def quantiles_to_outputs(quantiles:Iterable[float]):
             output_names.append('-median')
     return quantiles, output_names
 
-# %% ../nbs/src/utils.ipynb 37
+# %% ../nbs/src/utils.ipynb 38
 # given input array of sample forecasts and inptut quantiles/levels, 
 # output a Pandas Dataframe with columns of quantile predictions
 def samples_to_quantiles_df(samples: np.ndarray, 
@@ -557,3 +565,196 @@ def samples_to_quantiles_df(samples: np.ndarray,
                       columns=col_names)
     
     return _quantiles, pd.concat([data,df], axis=1).set_index('unique_id')
+
+# %% ../nbs/src/utils.ipynb 44
+# Masked empirical covariance matrix
+@njit("Array(float64, 2, 'F')(Array(float64, 2, 'C'), Array(bool, 2, 'C'))", nogil=NUMBA_NOGIL, cache=NUMBA_CACHE, parallel=NUMBA_PARALLEL, fastmath=NUMBA_FASTMATH, error_model="numpy")
+# @njit(nogil=NOGIL, cache=CACHE, parallel=True, fastmath=True, error_model="numpy")
+def _ma_cov(residuals: np.ndarray, not_nan_mask: np.ndarray):
+    """Masked empirical covariance matrix.
+
+    :meta private:
+    """
+    n_timeseries = residuals.shape[0]
+    W = np.zeros((n_timeseries, n_timeseries), dtype=np.float64).T
+    for i in prange(n_timeseries):
+        not_nan_mask_i = not_nan_mask[i]
+        for j in range(i + 1):
+            not_nan_mask_j = not_nan_mask[j]
+            not_nan_mask_ij = not_nan_mask_i & not_nan_mask_j   
+            n_samples = np.sum(not_nan_mask_ij)
+            # Only compute if we have enough non-nan samples in the time series pair
+            if n_samples > 1:
+                # Masked residuals
+                residuals_i = residuals[i][not_nan_mask_ij]
+                residuals_j = residuals[j][not_nan_mask_ij]
+                residuals_i_mean = np.mean(residuals_i)
+                residuals_j_mean = np.mean(residuals_j)
+                X_i = (residuals_i - residuals_i_mean)
+                X_j = (residuals_j - residuals_j_mean)
+                # Empirical covariance
+                factor_emp_cov = np.float64(1 / (n_samples - 1))
+                W[i, j] = W[j, i] = factor_emp_cov * np.sum(X_i * X_j)
+
+    return W
+
+# %% ../nbs/src/utils.ipynb 45
+# Shrunk covariance matrix using the Schafer-Strimmer method
+
+@njit("Array(float64, 2, 'F')(Array(float64, 2, 'C'), float64)", nogil=NUMBA_NOGIL, cache=NUMBA_CACHE, parallel=NUMBA_PARALLEL, fastmath=NUMBA_FASTMATH, error_model="numpy")
+def _shrunk_covariance_schaferstrimmer_no_nans(residuals: np.ndarray, mint_shr_ridge: float):
+    """Shrink empirical covariance according to the following method:
+        Schäfer, Juliane, and Korbinian Strimmer. 
+        ‘A Shrinkage Approach to Large-Scale Covariance Matrix Estimation and 
+        Implications for Functional Genomics’. Statistical Applications in 
+        Genetics and Molecular Biology 4, no. 1 (14 January 2005). 
+        https://doi.org/10.2202/1544-6115.1175.
+
+    :meta private:
+    """
+    n_timeseries = residuals.shape[0]
+    n_samples = residuals.shape[1]
+    
+    # We need the empirical covariance, the off-diagonal sum of the variance of 
+    # the empirical correlation matrix and the off-diagonal sum of the squared 
+    # empirical correlation matrix.
+    W = np.zeros((n_timeseries, n_timeseries), dtype=np.float64).T
+    sum_var_emp_corr = np.float64(0.0)
+    sum_sq_emp_corr = np.float64(0.0)
+    factor_emp_cov = np.float64(1 / (n_samples - 1))
+    factor_shrinkage = np.float64(1 / (n_samples * (n_samples - 1)))
+    epsilon = np.float64(2e-8)
+    for i in prange(n_timeseries):
+        # Mean of the standardized residuals
+        X_i = residuals[i] - np.mean(residuals[i])
+        Xs_i = X_i / (np.std(residuals[i]) + epsilon)
+        Xs_i_mean = np.mean(Xs_i)
+        for j in range(i + 1):
+            # Empirical covariance
+            X_j = residuals[j] - np.mean(residuals[j])
+            W[i, j] = factor_emp_cov * np.sum(X_i * X_j)
+            # Off-diagonal sums
+            if i != j:
+                Xs_j = X_j / (np.std(residuals[j]) + epsilon)
+                Xs_j_mean = np.mean(Xs_j)
+                # Sum off-diagonal variance of empirical correlation
+                w = (Xs_i - Xs_i_mean) * (Xs_j - Xs_j_mean)
+                w_mean = np.mean(w)
+                sum_var_emp_corr += np.sum(np.square(w - w_mean))
+                # Sum squared empirical correlation
+                sum_sq_emp_corr += w_mean**2
+
+    # Calculate shrinkage intensity 
+    shrinkage = 1.0 - max(min((factor_shrinkage * sum_var_emp_corr) / (sum_sq_emp_corr + epsilon), 1.0), 0.0)
+    # Shrink the empirical covariance
+    for i in prange(n_timeseries):
+        for j in range(i + 1):
+            if i != j:    
+                W[i, j] = W[j, i] = shrinkage * W[i, j]
+            else:
+                W[i, j] = W[j, i] = max(W[i, j], mint_shr_ridge)
+    return W
+
+@njit("Array(float64, 2, 'F')(Array(float64, 2, 'C'), Array(bool, 2, 'C'), float64)", nogil=NUMBA_NOGIL, cache=NUMBA_CACHE, parallel=NUMBA_PARALLEL, fastmath=NUMBA_FASTMATH, error_model="numpy")
+def _shrunk_covariance_schaferstrimmer_with_nans(residuals: np.ndarray, not_nan_mask: np.ndarray, mint_shr_ridge: float):
+    """Shrink empirical covariance according to the following method:
+        Schäfer, Juliane, and Korbinian Strimmer. 
+        ‘A Shrinkage Approach to Large-Scale Covariance Matrix Estimation and 
+        Implications for Functional Genomics’. Statistical Applications in 
+        Genetics and Molecular Biology 4, no. 1 (14 January 2005). 
+        https://doi.org/10.2202/1544-6115.1175.
+
+    :meta private:
+    """
+    n_timeseries = residuals.shape[0]
+    
+    # We need the empirical covariance, the off-diagonal sum of the variance of 
+    # the empirical correlation matrix and the off-diagonal sum of the squared 
+    # empirical correlation matrix.
+    W = np.zeros((n_timeseries, n_timeseries), dtype=np.float64).T
+    sum_var_emp_corr = np.float64(0.0)
+    sum_sq_emp_corr = np.float64(0.0)
+    epsilon = np.float64(2e-8)
+    for i in prange(n_timeseries):
+        not_nan_mask_i = not_nan_mask[i]
+        for j in range(i + 1):
+            not_nan_mask_j = not_nan_mask[j]
+            not_nan_mask_ij = not_nan_mask_i & not_nan_mask_j   
+            n_samples = np.sum(not_nan_mask_ij)
+            # Only compute if we have enough non-nan samples in the time series pair
+            if n_samples > 1:
+                # Masked residuals
+                residuals_i = residuals[i][not_nan_mask_ij]
+                residuals_j = residuals[j][not_nan_mask_ij]
+                residuals_i_mean = np.mean(residuals_i)
+                residuals_j_mean = np.mean(residuals_j)
+                X_i = (residuals_i - residuals_i_mean)
+                X_j = (residuals_j - residuals_j_mean)
+                # Empirical covariance
+                factor_emp_cov = np.float64(1 / (n_samples - 1))
+                W[i, j] = factor_emp_cov * np.sum(X_i * X_j)
+                # Off-diagonal sums
+                if i != j:
+                    factor_var_emp_cor = np.float64(n_samples / (n_samples - 1)**3)
+                    residuals_i_std = np.std(residuals_i) + epsilon
+                    residuals_j_std = np.std(residuals_j) + epsilon
+                    Xs_i = X_i / (residuals_i_std + epsilon)
+                    Xs_j = X_j / (residuals_j_std + epsilon)
+                    Xs_im_mean = np.mean(Xs_i)
+                    Xs_jm_mean = np.mean(Xs_j)
+                    # Sum off-diagonal variance of empirical correlation
+                    w = (Xs_i - Xs_im_mean) * (Xs_j - Xs_jm_mean)
+                    w_mean = np.mean(w)
+                    sum_var_emp_corr += factor_var_emp_cor * np.sum(np.square(w - w_mean))
+                    # Sum squared empirical correlation
+                    sum_sq_emp_corr += np.square(factor_emp_cov * n_samples * w_mean)
+
+    # Calculate shrinkage intensity 
+    shrinkage = 1.0 - max(min((sum_var_emp_corr) / (sum_sq_emp_corr + epsilon), 1.0), 0.0)
+
+    # Shrink the empirical covariance
+    for i in prange(n_timeseries):
+        for j in range(i + 1):
+            if i != j:    
+                W[i, j] = W[j, i] = shrinkage * W[i, j]
+            else:
+                W[i, j] = W[j, i] = max(W[i, j], mint_shr_ridge)
+
+    return W
+
+# %% ../nbs/src/utils.ipynb 47
+# Lasso cyclic coordinate descent
+@njit("Array(float64, 1, 'C')(Array(float64, 2, 'C'), Array(float64, 1, 'C'), float64, int64, float64)", nogil=NUMBA_NOGIL, cache=NUMBA_CACHE, parallel=True, fastmath=True, error_model="numpy")
+def _lasso(X: np.ndarray, y: np.ndarray, 
+          lambda_reg: float, max_iters: int = 1_000,
+          tol: float = 1e-4):
+    # lasso cyclic coordinate descent
+    n, feats = X.shape
+    norms = (X ** 2).sum(axis=0)
+    beta = np.zeros(feats, dtype=np.float64)
+    beta_changes = np.zeros(feats, dtype=np.float64)
+    residuals = y.copy()
+    
+    for it in range(max_iters):
+        for i, betai in enumerate(beta):
+            # is feature is close to zero, we 
+            # continue to the next.
+            # in this case is optimal betai= 0
+            if abs(norms[i]) < 1e-8:
+                continue
+            xi = X[:, i]
+            #we calculate the normalized derivative
+            # rho = betai
+            # for j in range(n):
+            #     rho += xi[j] * residuals[j] / norms[i]
+            # rho = betai + np.dot(xi, residuals) / norms[i] #(norms[i] + 1e-3)
+            rho = betai + xi.flatten().dot(residuals) / norms[i] #(norms[i] + 1e-3)
+            #soft threshold
+            beta[i] = np.sign(rho) * max(np.abs(rho) - lambda_reg * n / norms[i], 0.)#(norms[i] + 1e-3), 0.)
+            beta_changes[i] = np.abs(betai - beta[i])
+            if beta[i] != betai:
+                residuals += (betai - beta[i]) * xi
+        if max(beta_changes) < tol:
+            break
+
+    return beta
