+++ tests/test_benchmark.py
@@ -1,19 +1,22 @@
 
 import numpy as np
 import pytest
-import pytest_benchmark
 from scipy import sparse
-from hierarchicalforecast.methods import MinTrace, _ma_cov
+from hierarchicalforecast.methods import MinTrace, ERM
+from hierarchicalforecast.utils import _ma_cov
+
+import pytest_benchmark # noqa: F401
+
 #%% MinT benchmarks
 # run using: pytest tests\test_benchmark.py -v -s --benchmark-min-rounds=20
-def _create_mint_inputs(n_bottom_timeseries):
+def _create_reconciler_inputs(n_bottom_timeseries):
     # Create random hierarchy
     h = 100
     insample = 1000
     max_levels_random = 5
     max_categories_per_random_level = 10
     rng = np.random.default_rng(0)
-    ones = np.ones(n_bottom_timeseries, dtype=np.float32)
+    ones = np.ones(n_bottom_timeseries, dtype=np.float64)
     idx_range = np.arange(n_bottom_timeseries)
     n_levels_random = rng.integers(1, max_levels_random + 1)
     S_aggs_list = []
@@ -24,13 +27,13 @@ def _create_mint_inputs(n_bottom_timeseries):
         S_aggs_list.append(S_agg)
     S_aggs = sparse.vstack(S_aggs_list)
     # Create top and bottom level
-    S_top = sparse.csr_matrix(ones, dtype=np.float32)
-    S_bottom = sparse.eye(n_bottom_timeseries, dtype=np.float32, format="csr")
+    S_top = sparse.csr_matrix(ones, dtype=np.float64)
+    S_bottom = sparse.eye(n_bottom_timeseries, dtype=np.float64, format="csr")
     # Construct S: stack top, aggregations and bottom 
     S_sp = sparse.vstack([S_top, S_aggs, S_bottom])
 
-    y_hat_bottom = np.vstack([i * np.ones(h, dtype=np.float32) for i in range(n_bottom_timeseries)])
-    y_hat_bottom_insample = np.vstack([i * np.ones(insample, dtype=np.float32) for i in range(n_bottom_timeseries)])
+    y_hat_bottom = np.vstack([i * np.ones(h, dtype=np.float64) for i in range(n_bottom_timeseries)])
+    y_hat_bottom_insample = np.vstack([i * np.ones(insample, dtype=np.float64) for i in range(n_bottom_timeseries)])
     y_bottom = y_hat_bottom_insample + rng.normal(size=(n_bottom_timeseries, insample))
 
     S = S_sp.toarray()
@@ -41,26 +44,35 @@ def _create_mint_inputs(n_bottom_timeseries):
 
     return S, y_hat, y_insample, y_hat_insample, idx_bottom
 
-@pytest.mark.parametrize("n_bottom_timeseries", [20, 200, 2000])
+@pytest.mark.parametrize("n_bottom_timeseries", [20])
 @pytest.mark.parametrize("with_nans", (False, True))
 def test_mint(benchmark, n_bottom_timeseries, with_nans):
-    S, y_hat, y_insample, y_hat_insample, idx_bottom = _create_mint_inputs(n_bottom_timeseries)
+    S, y_hat, y_insample, y_hat_insample, idx_bottom = _create_reconciler_inputs(n_bottom_timeseries)
+
     if with_nans:
         y_insample[-1, :-1] = np.nan
         y_hat_insample[-1, :-1] = np.nan
 
     cls_min_trace = MinTrace(method='mint_shrink')
-    result_min_trace = benchmark(cls_min_trace, S=S, y_hat=y_hat, y_insample=y_insample, y_hat_insample=y_hat_insample, idx_bottom=idx_bottom)
+    result_min_trace = benchmark(cls_min_trace, S=S, y_hat=y_hat, y_insample=y_insample, y_hat_insample=y_hat_insample, idx_bottom=idx_bottom) # noqa: F841
 
-@pytest.mark.parametrize("n_bottom_timeseries", [20, 200, 2000])
+@pytest.mark.parametrize("n_bottom_timeseries", [20])
 @pytest.mark.parametrize("with_nans", (False, True))
 def test_cov(benchmark, n_bottom_timeseries, with_nans):
-    S, y_hat, y_insample, y_hat_insample, idx_bottom = _create_mint_inputs(n_bottom_timeseries)
+    S, y_hat, y_insample, y_hat_insample, idx_bottom = _create_reconciler_inputs(n_bottom_timeseries)
     if with_nans:
         y_insample[-1, :-1] = np.nan
         y_hat_insample[-1, :-1] = np.nan
 
-    residuals = (y_insample - y_hat_insample).astype(np.float32)
+    residuals = (y_insample - y_hat_insample)
     nan_mask = np.isnan(residuals)
 
-    result = benchmark(_ma_cov, residuals, ~nan_mask)
+    result = benchmark(_ma_cov, residuals, ~nan_mask) # noqa: F841
+
+@pytest.mark.parametrize("n_bottom_timeseries", [10])
+@pytest.mark.parametrize("erm_method", ['reg', 'reg_bu'])
+def test_erm_reg(benchmark, n_bottom_timeseries, erm_method):
+    S, y_hat, y_insample, y_hat_insample, idx_bottom = _create_reconciler_inputs(n_bottom_timeseries)
+
+    cls_min_trace = ERM(method=erm_method)
+    result_min_trace = benchmark(cls_min_trace, S=S, y_hat=y_hat, y_insample=y_insample, y_hat_insample=y_hat_insample, idx_bottom=idx_bottom) # noqa: F841
